+++ src/main/java/com/jane/ecommerce/interfaces/dto/user/ChargeResponse.java
@@ -4,9 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class ChargeResponse {
-    int currentBalance;
+    BigDecimal currentBalance;
 }
